==> pipelines/etl/2.transform_financial.ipynb <==
import yaml
import pyodbc
from market_growth_analysis.etl.stagging import *
import pandas as pd

# Load the YAML file
with open('../../conf/local.yml', 'r') as f:
    config = yaml.safe_load(f)

balance_sheet_df = pd.read_csv("../../data/raw_01/balance-sheet.csv", index_col=0)
cash_flow_statement_df = pd.read_csv("../../data/raw_01/cash-flow-statement.csv", index_col=0)
# financial_ratios = pd.read_csv("../../data/raw_01/balance-sheet.csv")
income_statement_df = pd.read_csv("../../data/raw_01/income-statement.csv", index_col=0)
# prices_df = pd.read_csv("../../data/raw_01/prices.csv", index_col=0)

# convert 'Date' column to datetime format
balance_sheet_df['Date'] = pd.to_datetime(balance_sheet_df['Date'])
cash_flow_statement_df['Date'] = pd.to_datetime(cash_flow_statement_df['Date'])
income_statement_df['Date'] = pd.to_datetime(income_statement_df['Date'])


# Perform inner joins between the dataframes
merged_df = balance_sheet_df.merge(cash_flow_statement_df, on=['ticker', 'Date'], how='inner')
merged_df = merged_df.merge(income_statement_df, on=['ticker', 'Date'], how='inner')

# Check for duplicates

# extract year from the 'Date' column
merged_df['Year'] = merged_df['Date'].dt.year

# check for duplicates in the combination of 'ID_COMPANY' and 'Year' columns
duplicated_rows = merged_df.duplicated(subset=['ticker', 'Year'], keep=False)

# print the duplicated rows
merged_df[['ticker', 'Year', 'Date']].loc[duplicated_rows]

merged_df['ticker'].loc[duplicated_rows].unique()

merged_df['n_nan'] = merged_df.isna().sum(axis=1).astype(int)

# sort the dataframe by 'ID_COMPANY', 'Year' and the count of missing values
df_sorted = merged_df.sort_values(by=['ticker', 'Year', 'n_nan'])

# keep the first occurrence of each 'ID_COMPANY-Year' combination
df_unique = df_sorted.drop_duplicates(subset=['ticker', 'Year'], keep='first')

df_unique[df_unique['ticker']=='NUZE']

df_unique.to_csv("../../data/intermediate_02/financial-sheets.csv")
