==> pipelines/etl/5.clean_data.ipynb <==
import yaml
import pyodbc
from market_growth_analysis.etl.stagging import *
import pandas as pd
import numpy as np
from datetime import datetime
# Load the YAML file
with open('../../conf/global.yml', 'r') as f:
    columns = yaml.safe_load(f)

financial = pd.read_csv("../../data/processed_03/financial-prices-ratios.csv", index_col=0)

# Clean Ratios

columns_ratios = columns['columns_ratios']
columns_ratios_all = columns['column_id'] + columns['columns_ratios']
ratios_df = financial[columns_ratios_all]

for column in columns_ratios:
    alpha = 0.1
    top_percentile = 1 - (alpha/2)
    low_percentile = alpha/2

    # Calculate the alpha percentile value
    top_percentile_value = ratios_df[column].quantile(top_percentile)
    low_percentile_value = ratios_df[column].quantile(low_percentile)

    # Replace outliers with percentiles value
    ratios_df[column] = ratios_df[column].where(ratios_df[column] <= top_percentile_value, top_percentile_value)
    ratios_df[column] = ratios_df[column].where(ratios_df[column] >= low_percentile_value, low_percentile_value)

# # Replace -inf and inf with None
# ratios_df = ratios_df.replace([-np.inf, np.inf], None)

# Make sure types are correct
ratios_df[columns['columns_ratios']] = ratios_df[columns['columns_ratios']].astype(float)
ratios_df['PK'] = ratios_df['PK'].astype(int)
ratios_df['Date'] = pd.to_datetime(ratios_df['Date'])



ratios_df.describe()

ratios_df.head(2)

# Clean Income Statement

columns_income_statement = columns['column_id'] + columns['columns_income_statement']
income_statememnt_df = financial[columns_income_statement]

income_statememnt_df.describe()

# Make sure types are correct
income_statememnt_df[columns['columns_income_statement']] = income_statememnt_df[columns['columns_income_statement']].astype(float)
income_statememnt_df['PK'] = income_statememnt_df['PK'].astype(int)
income_statememnt_df['Date'] = pd.to_datetime(income_statememnt_df['Date'])

income_statememnt_df.describe()

# Clean balance sheet

columns_balance_sheet = columns['column_id'] + columns['columns_balance_sheet']
balance_sheet_df = financial[columns_balance_sheet]

# Make sure types are correct
balance_sheet_df[columns['columns_balance_sheet']] = balance_sheet_df[columns['columns_balance_sheet']].astype(float)
balance_sheet_df['PK'] = balance_sheet_df['PK'].astype(int)
balance_sheet_df['Date'] = pd.to_datetime(balance_sheet_df['Date'])

# Clean columns_cash_flow_statement

columns_cash_flow_statement = columns['column_id'] + columns['columns_cash_flow_statement']
cash_flow_statement_df = financial[columns_cash_flow_statement]

# Make sure types are correct
cash_flow_statement_df[columns['columns_cash_flow_statement']] = cash_flow_statement_df[columns['columns_cash_flow_statement']].astype(float)
cash_flow_statement_df['PK'] = cash_flow_statement_df['PK'].astype(int)
cash_flow_statement_df['Date'] = pd.to_datetime(cash_flow_statement_df['Date'])

# Clean prices

columns_prices = columns['column_id'] + columns['columns_prices']
prices_df = financial[columns_prices]

prices_df.describe()

# Make sure types are correct
prices_df[columns['columns_prices']] = prices_df[columns['columns_prices']].astype(float)
prices_df[['PK','longevity']] = prices_df[['PK','longevity']].astype(int)
prices_df['Date'] = pd.to_datetime(prices_df['Date'])

# Create Dim table

columns_dim = columns['columns_dim']
dim_df = financial[columns_dim]

dim_grouped_df = dim_df.groupby('ticker').first().reset_index()

# Save dataframes to load

ratios_df.to_csv("../../data/load_04/ratios_df.csv")
income_statememnt_df.to_csv("../../data/load_04/income_statememnt_df.csv")
balance_sheet_df.to_csv("../../data/load_04/balance_sheet_df.csv")
cash_flow_statement_df.to_csv("../../data/load_04/cash_flow_statement_df.csv")
prices_df.to_csv("../../data/load_04/prices_df.csv")
dim_grouped_df.to_csv("../../data/load_04/dim_grouped_df.csv")
